requirements.txt
# ============================================================================
# Sistema de Punto de Venta con IA - Bazar Gulpery
# Dependencias del Proyecto
# ============================================================================

# ==================== VISIÓN POR COMPUTADORA ====================
opencv-python==4.8.1.78
ultralytics==8.1.0
torch>=2.0.0
torchvision>=0.15.0

# ==================== BASE DE DATOS ====================
psycopg2-binary==2.9.9

# ==================== ASISTENTE DE VOZ ====================
SpeechRecognition==3.10.1
pyttsx3==2.90
PyAudio==0.2.14

# ==================== DATASET Y ENTRENAMIENTO ====================
roboflow==1.1.17

# ==================== UTILIDADES ====================
numpy>=1.24.0
pillow>=10.0.0

# ==================== DESARROLLO Y LOGGING ====================
python-dotenv==1.0.0
